refactor(cli): isolate memfs token reporting
Keep `memfs.ts` focused on dispatch and git-backed memory operations. The token estimator path now lives in `memfsTokens.ts`, which owns input resolution, validation, formatting, and exit codes.

This is just cleanup: no CLI behavior changes.

Verification:
- `bun run typecheck`
- `bun test src/tests/utils/systemPromptSize.test.ts src/tests/cli/memfs-tokens.test.ts src/tests/permissions/readOnlyShell.test.ts`
- `letta memfs tokens` smoke against this agent's memory dir
